Added note about key/resource name
Note: For our purposes, the key is defined as the name of the resource. There can only be one key/name of resource within an individual stack. For instance, if you have coreo_aws_ec2_instance "my-instance" do somewhere and again have coreo_aws_ec2_instance "my-instance" do close to the root, the root will win! And it will override the first instance. Simply put, the resource name is the key and vice versa.